Move createProvider to IModelUtils
